🧶 Refactor org cards.
