+++ locales_dev/en/translation.json
@@ -128,7 +128,7 @@
   "Syncing GitHub Repos…": "Syncing GitHub Repos…",
   "Task": "Task",
   "Task Name": "Task Name",
-  "Task Orgs": "Task Orgs",
+  "Task Team & Orgs": "Task Team & Orgs",
   "Tasks for": "Tasks for",
   "There are no collaborators on this project": "There are no collaborators on this project",
   "There is an existing Dev Org for this task.": "There is an existing Dev Org for this task.",
