+++ package.json
@@ -50,7 +50,7 @@
   "dependencies": {
     "@salesforce-ux/design-system": "2.11.2",
     "@salesforce/design-system-react": "0.10.17",
-    "@sentry/browser": "^5.11.1",
+    "@sentry/browser": "^5.11.2",
     "classnames": "^2.2.6",
     "date-fns": "^2.9.0",
     "i18next": "^19.1.0",
@@ -96,7 +96,7 @@
     "@types/react-redux": "^7.1.7",
     "@types/react-router-dom": "^5.1.3",
     "@types/redux-logger": "^3.0.7",
-    "@types/uuid": "^3.4.6",
+    "@types/uuid": "^3.4.7",
     "@typescript-eslint/eslint-plugin": "^2.18.0",
     "@typescript-eslint/parser": "^2.18.0",
     "autoprefixer": "^9.7.4",
