+++ src/js/components/tasks/table.tsx
@@ -12,8 +12,8 @@ import {
   AssignUserModal,
   GitHubUserAvatar,
 } from '@/components/user/githubUser';
-import { GitHubUser } from '@/store/repositories/reducer';
 import { Task } from '@/store/tasks/reducer';
+import { GitHubUser } from '@/store/user/reducer';
 import { ORG_TYPES, OrgTypes, TASK_STATUSES } from '@/utils/constants';
 import routes from '@/utils/routes';
