+++ src/js/store/repositories/reducer.ts
@@ -1,13 +1,8 @@
 import { ObjectsAction, PaginatedObjectResponse } from '@/store/actions';
 import { RepositoriesAction } from '@/store/repositories/actions';
 import { LogoutAction } from '@/store/user/actions';
+import { GitHubUser } from '@/store/user/reducer';
 import { OBJECT_TYPES } from '@/utils/constants';
-
-export interface GitHubUser {
-  id: string;
-  login: string;
-  avatar_url: string;
-}
 
 export interface Repository {
   id: string;
